Serverspec checks for issue #46
This establishes a check for issue #46 via serverspec and completes the list of tests which had been implemented in BATS (at least for redhat-based systems).

I believe this also improves on the bats-based test by checking the contents of the /etc/exports file to ensure that chaining of anonuid/anongid hasn't snuck in.